+++ lib/gui/app/app.js
@@ -98,7 +98,6 @@ const app = angular.module('Etcher', [
 
   // OS
   require('./os/open-external/open-external'),
-  require('./os/dropzone/dropzone'),
 
   // Utils
   require('./utils/manifest-bind/manifest-bind')
