+++ lib/gui/app/components/image-selector/index.js
@@ -16,6 +16,8 @@
 
 'use strict'
 
+/* eslint-disable jsdoc/require-example */
+
 /**
  * @module Etcher.Components.ImageSelector
  */
@@ -24,11 +26,15 @@ const angular = require('angular')
 const { react2angular } = require('react2angular')
 
 const MODULE_NAME = 'Etcher.Components.ImageSelector'
-const SelectImageButton = angular.module(MODULE_NAME, [])
+const ImageSelector = angular.module(MODULE_NAME, [])
 
-SelectImageButton.component(
+ImageSelector.component(
   'imageSelector',
-  react2angular(require('./image-selector.jsx'))
+  react2angular(require('./image-selector.jsx')),
+  [],
+  [
+    'FileSelectorService',
+    'WarningModalService'
+  ]
 )
-
 module.exports = MODULE_NAME
