+++ lib/gui/app/pages/main/main.js
@@ -47,14 +47,12 @@ const MainPage = angular.module(MODULE_NAME, [
   require('../../components/drive-selector'),
 
   require('../../os/open-external/open-external'),
-  require('../../os/dropzone/dropzone'),
 
   require('../../utils/byte-size/byte-size'),
   require('../../utils/middle-ellipsis/filter')
 ])
 
 MainPage.controller('MainController', require('./controllers/main'))
-MainPage.controller('ImageSelectionController', require('./controllers/image-selection'))
 MainPage.controller('DriveSelectionController', require('./controllers/drive-selection'))
 MainPage.controller('FlashController', require('./controllers/flash'))
